Fix filtering on unique records by breaking
Before, the code would write the sdblock on every matched record
which could lead to a block to be written twice.
Instead, we break if we hit a value we've seen before.
We store all of the values in the matched record, so if we have
several records with the same field name, we keep all of them.
If we haven't broken, then we write.